pyproject.toml
[tool.poetry]
name = "dobby"
version = "0.1.0"
description = "Dobby is a CI/CD helper to deploy templated (jinja2) tasks to Nomad"
authors = ["Florian Apolloner <florian@apolloner.eu>"]
license = "BSD-3-Clause"

homepage = "https://github.com/apollo13/dobby"
repository = "https://github.com/apollo13/dobby"

classifiers = [
    "License :: OSI Approved :: BSD License",
]

[tool.poetry.scripts]
dobby = 'dobby.cli:cli'

[tool.poetry.dependencies]
python = ">3.7,<3.9"
jinja2 = "^2.11.1"
httpx = "^0.12.0"
click = "^7.1.1"
colorama = "^0.4.3"
PyYAML = "^5.3.1"
importlib_metadata = { version = "^1.5.0", python = "<3.8" }
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.7.9"
pex = "2.1.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
